LayoutTests/platform/ios-sim/fast/dom/navigator-vibration-expected.txt

Test for the Vibration API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL navigator.webkitVibrate(0); should be undefined. Threw exception TypeError: 'undefined' is not a function (evaluating 'navigator.webkitVibrate(0)')
FAIL navigator.webkitVibrate([]); should be undefined. Threw exception TypeError: 'undefined' is not a function (evaluating 'navigator.webkitVibrate([])')
FAIL navigator.webkitVibrate(1000); should be undefined. Threw exception TypeError: 'undefined' is not a function (evaluating 'navigator.webkitVibrate(1000)')
FAIL navigator.webkitVibrate([1000, 300, 500]); should be undefined. Threw exception TypeError: 'undefined' is not a function (evaluating 'navigator.webkitVibrate([1000, 300, 500])')
FAIL navigator.webkitVibrate(4294967295); should be undefined. Threw exception TypeError: 'undefined' is not a function (evaluating 'navigator.webkitVibrate(4294967295)')
PASS navigator.webkitVibrate(); threw exception TypeError: 'undefined' is not a function (evaluating 'navigator.webkitVibrate()').
PASS successfullyParsed is true

TEST COMPLETE
